Fixed issues/3
The Python 3 patch broke writing to standard output. This fixes the problem by
moving the call to pconv.convert() back to its proper place, and by accounting
for the different approaches to encoded output on sys.stdout in Python 2 and
Python 3 when opening the `out` filehandle for the first time.

Bug report: https://github.com/jgrassler/mkdocs-pandoc/issues/3